% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong_data_misc.R
\docType{data}
\name{abstracts}
\alias{abstracts}
\alias{abstracts_dfm}
\title{Abstracts of communication journals dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2500 rows and 1 columns.}
\usage{
abstracts

abstracts_dfm
}
\description{
This is a random sample of all abstracts of papers published in high-impact communication journals from 2000 to 2017. The one ends with "dfm" is same data in quanteda::dfm (document-feature matrix) format.
}
\keyword{datasets}
