% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong_data_misc.R
\docType{data}
\name{abstracts_seededlda}
\alias{abstracts_seededlda}
\alias{abstracts_btm}
\title{Topic models trained with the abstracts dataset.}
\format{
An object of class \code{textmodel_lda} (inherits from \code{textmodel}, \code{list}) of length 10.

An object of class \code{BTM} of length 9.
}
\usage{
abstracts_seededlda

abstracts_btm
}
\description{
These are topic models trained with different topic model packages.
}
\keyword{datasets}
