% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{abstracts_stm}
\alias{abstracts_stm}
\alias{abstracts_warplda}
\alias{abstracts_topicmodels}
\title{Topic models trained with the abstracts dataset.}
\format{An object of class \code{STM} of length 11.}
\usage{
abstracts_stm

abstracts_warplda

abstracts_topicmodels
}
\description{
These are topic models trained with different topic model packages.
}
\keyword{datasets}
