% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{afinn}
\alias{afinn}
\title{AFINN dictionary}
\format{An object of class \code{dictionary2} of length 11.}
\usage{
afinn
}
\description{
This is the AFINN sentiment dictionary in quanteda::dictionary format.
}
\keyword{datasets}
