% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong_update.R
\name{check_oolong}
\alias{check_oolong}
\title{Check whether the oolong needs to be updated}
\usage{
check_oolong(oolong, verbose = FALSE)
}
\arguments{
\item{oolong}{an oolong object to be checked}

\item{verbose, }{logical, display messages}
}
\value{
Nothing
}
\description{
This function raises an error when the input oolong object needs to be updated. Oolong objects generated with an old version of oolong need to be updated to use the functionalities from the recent versions of oolong.
}
\author{
Chung-hong Chan
}
