% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong_summary_tm.R
\name{clone_oolong}
\alias{clone_oolong}
\title{Clone a oolong object.}
\usage{
clone_oolong(oolong)
}
\arguments{
\item{oolong}{an oolong object.}
}
\description{
Clone a new oolong object. The oolong must not be locked and ever coded.
}
