% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong.R
\name{create_oolong}
\alias{create_oolong}
\title{Generate a oolong test for a topic model}
\usage{
create_oolong(model, n_top_terms = 5, bottom_terms_percentile = 0.6,
  difficulty = 0.8)
}
\arguments{
\item{model}{a STM or WrapLDA object}

\item{n_top_terms}{integer, number of top topic words to be included in the candidates}

\item{difficulty}{double, to adjust the difficulty of the test. Higher value indicates higher difficulty, must be within the range of 0 to 1.}

\item{bottm_terms_percentile}{double, a term is considered to be an instruder when its theta less than the percentile of this theta, must be within the range of 0 to 1.}
}
\description{
Currently, this function generates a oolong test object that contains a word instrusion test. Future version will provide additional tests such as topic instrusion test.
}
