% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong_shiny.R
\name{export_oolong}
\alias{export_oolong}
\title{Export a deployable Shiny app from an oolong object into a directory
This function transforms your oolong test into a deployable Shiny app that is in a directory. The Shiny app is both launchable with shiny::runApp() and deployable with rsconnect::deployApp(). See \code{deploy_oolong} and the deployment guide for more details.}
\usage{
export_oolong(oolong, dir = base::tempdir(), verbose = TRUE)
}
\arguments{
\item{oolong}{an oolong object to be exported. Please note that the "witi" type, i.e. oolong object with both word and topic intrusion tests, cannot be exported. Also the object must not be locked and ever coded.}

\item{dir}{character string, the directory to be exported. Default to a temporary directory}

\item{verbose}{logical, whether to display information after exporting}
}
\value{
directory exported, invisible
}
\description{
Export a deployable Shiny app from an oolong object into a directory
This function transforms your oolong test into a deployable Shiny app that is in a directory. The Shiny app is both launchable with shiny::runApp() and deployable with rsconnect::deployApp(). See \code{deploy_oolong} and the deployment guide for more details.
}
\examples{
# Please try this example in interactive R sessions only.
if (interactive()) {
   data(abstracts_stm)
   x <- wi(abstracts_stm)
   export_oolong(x)
}
}
\author{
Chung-hong Chan
}
