% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong.R
\docType{data}
\name{newsgroup5}
\alias{newsgroup5}
\title{Newsgroup text dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4182 rows and 3 columns.}
\usage{
newsgroup5
}
\description{
This is a subset of the famous "newsgroup 20" dataset.
}
\keyword{datasets}
