% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong_data_misc.R
\docType{data}
\name{newsgroup_nb}
\alias{newsgroup_nb}
\title{Naive Bayes model trained on 20 newsgroups data}
\format{
An object of class \code{textmodel_nb} (inherits from \code{textmodel}, \code{list}) of length 7.
}
\usage{
newsgroup_nb
}
\description{
This is a Naive Bayes model (of the class 'textmodel_nb') trained on 20 newsgroups data.
}
\references{
Lang, K. (1995). Newsweeder: Learning to filter netnews. In Machine Learning Proceedings 1995 (pp. 331-339). Morgan Kaufmann.
}
\keyword{datasets}
