% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong.R
\docType{package}
\name{oolong}
\alias{oolong}
\title{Oolong: create and administrate validation tests for typical automated content analysis tools}
\description{
The oolong package makes it very easy to create, administrate and evaluate typical automated content analysis tools by providing a framework for creating human-in-the-loop validation tests. For topic models, oolong can generate tests such as word intrusion test and topic intrusion test (Chang et al.). For dictionary-based methods, oolong can generate standardized interface for making gold standard ('Ground truth') data. There are only two core functions of this package: \code{\link{create_oolong}} and \code{\link{summarize_oolong}}.
}
\author{
Chung-hong Chan
}
