% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong_gs.R
\name{print.oolong_gold_standard}
\alias{print.oolong_gold_standard}
\title{Print oolong gold standard object}
\usage{
\method{print}{oolong_gold_standard}(x, ...)
}
\arguments{
\item{x}{an oolong gold standard object}

\item{...}{other parameters}
}
\value{
None, a summary of the quanteda::corpus and what you should do are displayed
}
\description{
This function prints a summary of the oolong gold standard object. An oolong gold standard object is a result of $turn_gold() method. It is a quanteda::corpus compatible object.
}
\author{
Chung-hong Chan
}
