% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong.R
\name{revert_oolong}
\alias{revert_oolong}
\title{Obtain a locked oolong from a downloaded data file}
\usage{
revert_oolong(oolong, rds_file)
}
\arguments{
\item{oolong}{an oolong object used for deployment}

\item{rds_file}{path to the downloaded RDS file}
}
\value{
a locked oolong object based on the data in the downloaded RDS file
}
\description{
To generate a locked oolong object with the original oolong object and the RDS file. The RDS file should have been downloaded from a deployed Shiny app.
}
\author{
Chung-hong Chan
}
