% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong_summary.R
\name{summarise_oolong}
\alias{summarise_oolong}
\alias{summarize_oolong}
\title{Summarize oolong objects}
\usage{
summarise_oolong(...)

summarize_oolong(..., target_value = NULL)
}
\arguments{
\item{...}{(tm/gs) one or more oolong objects to be summarized.}

\item{target_value}{(gs) a vector of numeric values, the value you want to validate against the human-coded gold standard. One example of this target value is sentiment score extracted automatically from text.}
}
\value{
An oolong summary
}
\description{
This function summarizes one or more oolong objects. All oolong objects must be locked.
}
\examples{
\dontrun{
data(abstracts_stm)
oolong_test1 <- create_oolong(abstracts_stm)
oolong_test2 <- clone_oolong(oolong_test1)
oolong_test1$do_word_intrusion_test()
oolong_test2$do_word_intrusion_test()
oolong_test1$lock()
oolong_test2$lock()
summarize(oolong_test1, oolong_test2)
}
}
