% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong_data_misc.R
\docType{data}
\name{trump2k}
\alias{trump2k}
\title{Trump's tweets dataset}
\format{
An object of class \code{character} of length 2000.
}
\usage{
trump2k
}
\description{
This is a random sample of 2000 tweets from @realdonaldtrump account before his assumption of duty as the president of the United States.
}
\keyword{datasets}
