% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong_update.R
\name{update_oolong}
\alias{update_oolong}
\title{Update an oolong object to the latest version}
\usage{
update_oolong(oolong, verbose = TRUE)
}
\arguments{
\item{oolong}{an oolong object to be updated}

\item{verbose, }{logical, display messages}
}
\value{
an updated oolong object
}
\description{
This function update an old oolong object to the latest version.
}
\author{
Chung-hong Chan
}
