% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong_data_misc.R
\docType{data}
\name{abstracts_stm}
\alias{abstracts_stm}
\alias{abstracts_warplda}
\alias{abstracts_btm}
\alias{abstracts_keyatm}
\title{Topic models trained with the abstracts dataset.}
\format{
An object of class \code{STM} of length 11.

An object of class \code{WarpLDA} (inherits from \code{LDA}, \code{TopicModel}, \code{mlapiDecomposition}, \code{mlapiTransformation}, \code{mlapiBase}, \code{R6}) of length 9.

An object of class \code{BTM} of length 9.

An object of class \code{keyATM_output} (inherits from \code{base}, \code{list}) of length 18.
}
\usage{
abstracts_stm

abstracts_warplda

abstracts_btm

abstracts_keyatm
}
\description{
These are topic models trained with different topic model packages.
}
\keyword{datasets}
