% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong_data_misc.R
\docType{data}
\name{abstracts_stm}
\alias{abstracts_stm}
\alias{abstracts_warplda}
\title{Topic models trained with the abstracts dataset.}
\format{An object of class \code{STM} of length 11.}
\usage{
abstracts_stm

abstracts_warplda
}
\description{
These are topic models trained with different topic model packages.
}
\keyword{datasets}
