% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong_data_misc.R
\docType{data}
\name{afinn}
\alias{afinn}
\title{AFINN dictionary}
\format{
An object of class \code{dictionary2} of length 11.
}
\usage{
afinn
}
\description{
This is the AFINN sentiment dictionary in quanteda::dictionary format.
}
\references{
Nielsen, F. Å. (2011). A new ANEW: Evaluation of a word list for sentiment analysis in microblogs. arXiv preprint arXiv:1103.2903.
}
\keyword{datasets}
