% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong.R
\name{clone_oolong}
\alias{clone_oolong}
\title{Clone an oolong object}
\usage{
clone_oolong(oolong)
}
\arguments{
\item{oolong}{an oolong object.}
}
\value{
an oolong object
}
\description{
Clone a new oolong object. The oolong must not be locked and ever coded.
}
\author{
Chung-hong Chan
}
