% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong.R
\name{create_oolong}
\alias{create_oolong}
\title{Generate a oolong test}
\usage{
create_oolong(input_model = NULL, input_corpus = NULL,
  n_top_terms = 5, bottom_terms_percentile = 0.6, exact_n = NULL,
  frac = 0.01, n_top_topics = 3, n_topiclabel_words = 8,
  use_frex_words = FALSE, difficulty = 1, input_dfm = NULL,
  target_value = NULL, construct = "positive")
}
\arguments{
\item{input_model}{a STM, WarpLDA or topicmodels object}

\item{input_corpus}{if input_model is not null, it should be the corpus (character vector or quanteda::corpus object) to generate the model object, if not NULL, topic intrusion test cases are generated. If input_model is null, it generates gold standard test cases.}

\item{n_top_terms}{integer, number of top topic words to be included in the candidates of word intrusion test}

\item{exact_n}{integer, number of topic intrusion test cases to generate, ignore if frac is not NULL}

\item{frac}{double, fraction of test cases to be generated from the corpus}

\item{n_topiclabel_words}{integer, number of topic words to be shown as the topic label}

\item{use_frex_words}{logical, for a STM object, use FREX words if TRUE, use PROB words if FALSE}

\item{difficulty}{double, to adjust the difficulty of the test. Higher value indicates higher difficulty, must be within the range of 0 to 1, no effect for STM if use_frex_words is FALSE. Ignore for topicmodels objects.}

\item{input_dfm}{a dfm object used for training the input_model, if input_model is a WarpLDA object}

\item{bottm_terms_percentile}{double, a term is considered to be an word intruder when its theta less than the percentile of this theta, must be within the range of 0 to 1}

\item{n_top_topic}{integer, number of most relevant topics to be shown alongside the intruder topic}
}
\description{
This function generates a oolong test object that can either be used for validating a topic model or for creating ground truth (gold standard) of a text corpus.
}
