% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong.R
\name{create_oolong}
\alias{create_oolong}
\title{Generate a oolong test for a topic model}
\usage{
create_oolong(input_model, input_corpus = NULL, n_top_terms = 5,
  bottom_terms_percentile = 0.6, exact_n = 15, frac = NULL,
  n_top_topics = 3, n_topiclabel_words = 8, difficulty = 0.8)
}
\arguments{
\item{input_model}{a STM or WrapLDA object}

\item{input_corpus}{the corpus (character vector or quanteda::corpus object) to generate the model object, if not NULL, topic intrusion test cases are generated}

\item{n_top_terms}{integer, number of top topic words to be included in the candidates of word intrusion test}

\item{exact_n}{integer, number of topic intrusion test cases to generate, ignore if frac is not NULL}

\item{frac}{double, fraction of test cases to be generated from the corpus}

\item{n_topiclabel_words}{integer, number of topic words to be shown as the topic label}

\item{difficulty}{double, to adjust the difficulty of the test. Higher value indicates higher difficulty, must be within the range of 0 to 1}

\item{bottm_terms_percentile}{double, a term is considered to be an word intruder when its theta less than the percentile of this theta, must be within the range of 0 to 1}

\item{n_top_topic}{integer, number of most relevant topics to be shown alongside the intruder topic}
}
\description{
Currently, this function generates a oolong test object that contains a word intrusion test. Future version will provide additional tests such as topic intrusion test.
}
