% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong_shiny.R
\name{deploy_oolong}
\alias{deploy_oolong}
\title{Deploy an oolong test}
\usage{
deploy_oolong(oolong)
}
\arguments{
\item{oolong}{an oolong object to be deployed. Please note that the "witi" type, i.e. oolong object with both word and topic intrusion tests, cannot be deployed. Also the object must not be locked and ever coded.}
}
\value{
Nothing, it launches a deployable version of the coding interface
}
\description{
In most of the time, you should not use this function. You should write the deployable version of your app into a directory using \code{export_oolong} instead. Please refer to \code{vignette("deploy", package = "oolong")} for more details.
}
\examples{
# Please try this example in interactive R sessions only.
if (interactive()) {
   data(abstracts_stm)
   x <- wi(abstracts_stm)
   deploy_oolong(x)
}
}
\author{
Chung-hong Chan
}
