% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong.R
\docType{package}
\name{oolong}
\alias{oolong}
\title{Oolong: create and administrate validation tests for typical automated content analysis tools}
\description{
The oolong package makes it very easy to create, administrate and evaluate typical automated content analysis tools by providing a framework for creating human-in-the-loop validation tests. For topic models, oolong can generate tests such as word intrusion test, topic intrusion test (Chang et al.) and word set intrusion test (Ying et al.) For dictionary-based methods, oolong can generate standardized interface for making gold standard ('Ground truth') data.
}
\author{
Chung-hong Chan
}
