% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolong_summary.R
\name{print.oolong_summary}
\alias{print.oolong_summary}
\alias{plot.oolong_summary}
\title{Print and plot oolong summary}
\usage{
\method{print}{oolong_summary}(x, ...)

\method{plot}{oolong_summary}(x, ...)
}
\arguments{
\item{x}{an oolong_summary}

\item{...}{other parameters}
}
\description{
These functions print or plot a useful summary of the results from \code{\link{summarize_oolong}}.
}
